Crate hello_race

One of the simplest possible race conditions. Exits with a nonzero code under one order, and not the other.

However, we complicate things a bit by having an rdtsc and a branch right before
the critical event, to make it easier to recognize in the raw schedule recordings,
as when manually checking the results of this test.

 [dependencies]
 core = "1.6.0"

Constants

WORK_AMT 🔒

Functions

do_work 🔒

main 🔒

rdtsc 🔒

run_test 🔒

thread1 🔒

thread2 🔒